Moves session validation to server components for dashboard
Improves security and performance by shifting authentication checks from client-side and middleware to server-side code.
Simplifies middleware to only check for session cookies, deferring actual session validation to server components.
Splits dashboard logic into server and client components for cleaner separation.
Removes unnecessary webpack config for external packages.
